Feature/cpp refactoring (#144)
* c++ refactoring

refactored the code based on tips and hints of the editor

* committing before linting

* updated propagation Ids

* more refactoring

* clang-format

* fixed issue in main

+++ src/search/cost.cpp
@@ -9,8 +9,8 @@ Cost::Cost(Int violationDegree, Int objective,
       _objectiveWeightSign(static_cast<int>(direction)) {}
 
 Int Cost::evaluate(UInt violationWeight, UInt objectiveWeight) const noexcept {
-  return violationWeight * _violationDegree +
-         objectiveWeight * _objectiveWeightSign * _objective;
+  return static_cast<Int>(violationWeight) * _violationDegree +
+         static_cast<Int>(objectiveWeight) * _objectiveWeightSign * _objective;
 }
 
 std::string Cost::toString() const {
